// src/api/openapi/auto-generated/models/DisclosureAvailability.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const DisclosureAvailability = {
    available: 'available',
    riskDriver: 'risk_driver',
    notAvailable: 'not_available',
    unknownDefaultOpenApi: '11184809'
} as const;
export type DisclosureAvailability = typeof DisclosureAvailability[keyof typeof DisclosureAvailability];


export function instanceOfDisclosureAvailability(value: any): boolean {
    for (const key in DisclosureAvailability) {
        if (Object.prototype.hasOwnProperty.call(DisclosureAvailability, key)) {
            if ((DisclosureAvailability as Record<string, DisclosureAvailability>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function DisclosureAvailabilityFromJSON(json: any): DisclosureAvailability {
    return DisclosureAvailabilityFromJSONTyped(json, false);
}

export function DisclosureAvailabilityFromJSONTyped(json: any, ignoreDiscriminator: boolean): DisclosureAvailability {
    return json as DisclosureAvailability;
}

export function DisclosureAvailabilityToJSON(value?: DisclosureAvailability | null): any {
    return value as any;
}