// src/api/openapi/auto-generated/models/EUHazardMetadata.ts
export interface EUHazardMetadata {
    /**
     * 
     * @type {DisclosureType}
     * @memberof EUHazardMetadata
     */
    type: DisclosureType;
    /**
     * 
     * @type {boolean}
     * @memberof EUHazardMetadata
     */
    show?: boolean;
    /**
     * 
     * @type {DisclosureCategory}
     * @memberof EUHazardMetadata
     */
    category: DisclosureCategory;
    /**
     * 
     * @type {DisclosureAvailability}
     * @memberof EUHazardMetadata
     */
    availability: DisclosureAvailability;
    /**
     * 
     * @type {string}
     * @memberof EUHazardMetadata
     */
    title: string | null;
    /**
     * 
     * @type {string}
     * @memberof EUHazardMetadata
     */
    subtitle: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof EUHazardMetadata
     */
    overviewMetrics: Array<string>;
    /**
     * 
     * @type {Array<DisclosureMetricMetadata>}
     * @memberof EUHazardMetadata
     */
    metrics: Array<DisclosureMetricMetadata>;
}